requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
python-dotenv
pandas
openpyxl
flask-bcrypt
Werkzeug
fpdf2
psycopg2-binary
langchain
langchain-community
pypdf
faiss-cpu
sentence-transformers
Faker
easyocr
opencv-python-headless
